Order items grouped by section & order id
